Migrate from SQL to MongoDB on Azure (#27)
* Migrate to use mongo

* Add docker files

* Set up CI with Azure Pipelines

[skip ci]

* Update composer.json

* Update composer.lock

* Add pkg

* pkg

* remove

* Create Dockerfile

* pkg

* pkg

* pkg

* pkg

* pkg

* pkg

* docker

* Update Dockerfile

* Update Payment.php

* Add or update the App Service deployment workflow configuration from Azure Portal.

* Add htaccess file

* Update cd-production.yml

* Change php version

* Update cd-production.yml

* Update cd-production.yml

* Update cd-production.yml

* Update cd-production.yml

* Update mongo-db-docker_guides-test.yml

* Update index.php

* Zip artifact

* Update .htaccess

* Update .htaccess

* Update .htaccess

* Update .htaccess

* config

* move file

* Update .htaccess

* Update .htaccess

* remove

* Create web.config

* configure

* Add or update the App Service deployment workflow configuration from Azure Portal.

* php 7.4

* Update mongo-db-docker_guides-win-test.yml

* Update mongo-db-docker_guides-test.yml

* Create web.conf

* Create .htaccess

* Delete not used files

* Add back in Sentry context

* Fix date not showing in Mongo

* Update IncomeController.php

* SQL to mongo seeders

* Update database.php

* Update mongo-db-docker_guides-test.yml

* Fix events

* Create FixBankAccountSeeder.php

* Update DatabaseSeeder.php

* Fix roles for users

* Update mongo-db-docker_guides-test.yml

* Update mongo-db-docker_guides-test.yml

* Update mongo-db-docker_guides-test.yml

* Update mongo-db-docker_guides-test.yml

* Update mongo-db-docker_guides-test.yml

* Update mongo-db-docker_guides-test.yml

* Update mongo-db-docker_guides-test.yml

* Fix missing created_at & updated_at

* Update RedirectIfAuthenticated.php

* Fix dates

* Create FixBankTransactionsSeeder.php

* Assign API token on approval

* Fix date time not showing in input fields

* Handle for new ref_id

* Handle for new ref_id

* Refactor generating new ids

* Update README.md

* Delete cd-production.yml

* Don't deploy to Azure

* Init terraform

* Only master

* Create terraform.yml

* Update terraform

* Update terraform

* Update terraform.yml

* Update terraform.yml

* Update pipeline

* Update azure.yml

* Update azure.yml

* Delete terraform

* Update UsersController.php

* Fix object ids not being created

* Fix seeders

* Fix for Mongo ids

* Fix ref ids

* Fix export pages not loading data

* Fix date formatting on export pages

* Fix export pages without showing a code

* Fix export file name showing underfined

* Clean up

* Fix bools saving as 1/0 in db

* Fix views with boolean checks

+++ app/Http/Controllers/Admin/EventController.php
@@ -52,7 +52,7 @@ class EventController extends Controller
         // Format Payments ID
         $payments = $event->payments;
         foreach ($payments as $index => $payment) {
-            $payments[$index]->keyID = "p_" . $payment->id;
+            $payments[$index]->keyID = "p_" . $payment->ref_id;
             
             if($payment->is_cash) {
                 $payments[$index]->cash_only = $payment->amount;
@@ -66,7 +66,7 @@ class EventController extends Controller
         // Format Incomes ID/ref
         $incomes = $event->incomes;
         foreach ($incomes as $index => $income) {
-            $incomes[$index]->keyID = "i_" . $income->id;
+            $incomes[$index]->keyID = "i_" . $income->ref_id;
 
             if($income->is_cash) {
                 $incomes[$index]->cash_and_cheque = $income->amount;
@@ -151,8 +151,8 @@ class EventController extends Controller
         // Find event
         $event = Event::findOrFail($id);
 
-        // Delete only if no payments or incomes associated
-        if ($event->payments->count() == 0 || $event->incomes->count() == 0) {
+        // Delete only if no payments and no incomes associated
+        if ($event->payments->count() == 0 && $event->incomes->count() == 0) {
             $event->delete();
             
             $request->session()->flash('alert-success', $event->title . ' event has been deleted');
